+++ src/controller/controller_cliente.py
@@ -71,13 +71,29 @@ class Controller_Cliente:
         cpf = int(input("Insira o CPF do cliente a ser excluído: "))
 
         if not self.verifica_se_existe(oracle, cpf):
-            df_cliente = oracle.sqlToDataFrame(f"select id_cliente, nome_cliente, telefone from clientes where id_cliente = '{cpf}'")
-            oracle.write(f"delete from clientes where id_cliente = '{cpf}'")
 
-            cliente_excluido = Cliente(df_cliente.id_cliente.values[0], df_cliente.nome_cliente.values[0], df_cliente.telefone.values[0])
+            if self.verifica_se_existe_agenda(oracle, cpf):
+                
+                confirmation = str(input("Tem certeza que quer excluir esse cleinte? (Digite S para sim e N para não) "))
+                if confirmation.upper() == "S":
+                    df_cliente = oracle.sqlToDataFrame(f"select id_cliente, nome_cliente, telefone from clientes where id_cliente = '{cpf}'")
+                    oracle.write(f"delete from clientes where id_cliente = '{cpf}'")
 
-            print("Cliente Removido com sucesso")
-            print(cliente_excluido.to_string())
+                    cliente_excluido = Cliente(df_cliente.id_cliente.values[0], df_cliente.nome_cliente.values[0], df_cliente.telefone.values[0])
+
+                    print("Cliente Removido com sucesso")
+                    print(cliente_excluido.to_string())
+            
+            confirmation = str(input("O usuário tem registro na tabela agenda, digite S para excluir os registros da tabela agenda e N para voltar ao menu principal: "))
+            if confirmation.upper() == "S":
+                df_cliente = oracle.sqlToDataFrame(f"select id_cliente, nome_cliente, telefone from clientes where id_cliente = '{cpf}'")
+                oracle.write(f"delete from agenda where id_cliente = '{cpf}'")
+                oracle.write(f"delete from clientes where id_cliente = '{cpf}'")
+
+                cliente_excluido = Cliente(df_cliente.id_cliente.values[0], df_cliente.nome_cliente.values[0], df_cliente.telefone.values[0])
+
+                print("Cliente Removido com sucesso")
+                print(cliente_excluido.to_string())
         
         else:
             print(f"cliente {cpf} não existe")
@@ -87,3 +103,8 @@ class Controller_Cliente:
 
         df_cliente = oracle.sqlToDataFrame(f"select id_cliente, nome_cliente from clientes where id_cliente = '{cpf}'")
         return df_cliente.empty
+
+    def verifica_se_existe_agenda(self, oracle:OracleQueries, id_agenda:int=None) -> bool:
+
+        df_agenda = oracle.sqlToDataFrame(f"select id_agenda, id_cliente, id_lab from agenda where id_cliente = '{id_agenda}'")
+        return df_agenda.empty
